RC-update: fix file-sync (#982)
* refactor: decommission FileExtractionService v1

* Bug/assets 977 document indexing failures after pre release (#980)

* Fix unhandled rejection in process-queue, crashing the file-sync process
error isolation with logging (per asset / file)
Fix standardFontDataUrl warning
updated s3-sync script to allow local sync to MiniIO and usage of new database shema asets <-> files

* limit chunkBulkOperations size

* downgraded pdfjs again. Too risky. font-loading error avoidance works without version upgrade

* adds missing await

---------
