+++ apps/server-asset-sg/src/features/assets/files/file.service.ts
@@ -1,3 +1,4 @@
+import * as path from 'path';
 import {
   Asset,
   AssetFile,
@@ -26,6 +27,12 @@ import { FileS3Service, SaveFileS3Options } from '@/features/assets/files/file-s
 import { CreateFileData, FileIdentifier, FileRepo } from '@/features/assets/files/file.repo';
 import { PdfMetadataService } from '@/features/assets/files/pdf-metadata.service';
 import { mapAssetLanguagesToPrismaUpdate } from '@/features/assets/prisma-asset';
+
+// eval('require') bypasses webpack's static require analysis so the path is resolved at runtime by Node.js.
+
+const STANDARD_FONT_DATA_URL =
+  path.join(path.dirname((eval('require') as NodeRequire).resolve('pdfjs-dist/package.json')), 'standard_fonts') +
+  path.sep;
 
 @Injectable()
 export class FileService {
@@ -224,6 +231,7 @@ export class FileService {
     try {
       doc = await getDocument({
         url: presignedUrl,
+        standardFontDataUrl: STANDARD_FONT_DATA_URL,
         disableAutoFetch: true,
         disableStream: false,
       }).promise;
@@ -239,7 +247,7 @@ export class FileService {
         yield { pageNumber: i, text };
       }
     } finally {
-      doc?.destroy();
+      await doc?.destroy();
     }
   }
