+++ libs/client-shared/src/lib/components/pdf-viewer/pdf-viewer.service.ts
@@ -397,7 +397,6 @@ export class PdfViewerService implements OnDestroy {
 
   private async destroyPdfJsWorker() {
     await this.loadingTask?.destroy();
-    await this.pdfDoc?.destroy();
   }
 
   private getAuthorizationHeader(): Record<string, string> {
